fixed rust max size
